+++ src/screens/home/FeedDetailScreen.tsx
@@ -1,4 +1,5 @@
-import { useRef, useState } from 'react';
+import { Suspense, useRef, useState } from 'react';
+import { ErrorBoundary } from 'react-error-boundary';
 import { useTranslation } from 'react-i18next';
 import {
   Alert,
@@ -9,13 +10,15 @@ import {
   TouchableOpacity,
   View,
 } from 'react-native';
-import { RoomRepository } from '@/api';
+import { CommentRepository, feedKeys, FeedRepository, RoomRepository } from '@/api';
 import { CommentList, FeedItem, KeyboardLayout, Layout, Input } from '@/components';
 import { useFeedDetail } from '@/hooks';
 import { FeedStackParamList } from '@/navigation/navigationRef';
 import { useModalStore, useUserStore } from '@/store';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
+import { useSuspenseQueries } from '@tanstack/react-query';
 import { MoreVertical, Send } from 'lucide-react-native';
+import FeedSkeleton from '@/components/feed/FeedSkeleton';
 import { FeedOptionModal } from '@/components/modal/BottomOption/FeedOptionModal';
 import { ChatRequestModal } from '@/components/modal/Common/ChatRequestModal';
 
@@ -30,12 +33,32 @@ export default function FeedDetailScreen({ navigation, route }: FeedDetailScreen
   const handleModalClose = useModalStore((state) => state.handleClose);
   const isCertificated = useUserStore((state) => state.isCertificated);
   const [parentCommentId, setParentCommentId] = useState<number | null>(null);
-  const myUserId = useUserStore((state) => state.id);
 
-  const { feed, comments, isRefetching, handleFeedActions, handleCommentActions, handleRefresh } =
-    useFeedDetail({
-      feedId,
-    });
+  const handleRefresh = async () => {
+    await Promise.all([refetchFeed(), refetchComments()]);
+  };
+
+  const results = useSuspenseQueries({
+    queries: [
+      {
+        queryKey: feedKeys.detail(feedId),
+        queryFn: () => FeedRepository.get({ feedId }),
+      },
+      {
+        queryKey: ['feedComments', feedId],
+        queryFn: () => CommentRepository.getComments({ feedId }),
+      },
+    ],
+  });
+
+  const [
+    { data: feed, refetch: refetchFeed, isRefetching: isRefetchingFeed },
+    { data: comments, refetch: refetchComments },
+  ] = results;
+
+  const { handleFeedActions, handleCommentActions } = useFeedDetail({
+    feedId,
+  });
 
   const showFeedNotFoundAlert = () => {
     Alert.alert(
@@ -51,8 +74,8 @@ export default function FeedDetailScreen({ navigation, route }: FeedDetailScreen
     );
   };
 
-  const handleRoomCreate = async () => {
-    const data = await RoomRepository.create({ buddyId: feed.userId });
+  const handleRoomCreate = async (id: number) => {
+    const data = await RoomRepository.create({ buddyId: id });
     navigation.navigate('Chat', { screen: 'ChatRoom', params: { ...data } } as any);
   };
 
@@ -84,7 +107,6 @@ export default function FeedDetailScreen({ navigation, route }: FeedDetailScreen
     }
   };
 
-  if (!(feed || isRefetching)) return null;
   return (
     <>
       <Layout
@@ -123,28 +145,24 @@ export default function FeedDetailScreen({ navigation, route }: FeedDetailScreen
             className="mt-1"
             refreshControl={
               <RefreshControl
-                refreshing={isRefetching}
+                refreshing={isRefetchingFeed}
                 onRefresh={handleRefresh}
                 tintColor="#4AA366"
               />
             }
           >
-            {feed && (
-              <>
-                <FeedItem
-                  feed={feed}
-                  onLike={handleFeedActions.like}
-                  onBookmark={handleFeedActions.bookmark}
-                  showAllContent
-                />
-                <CommentList
-                  feed={feed}
-                  comments={comments}
-                  onReply={handleCommentReply}
-                  onLike={handleCommentActions.like}
-                />
-              </>
-            )}
+            <FeedItem
+              feed={feed}
+              onLike={handleFeedActions.like}
+              onBookmark={handleFeedActions.bookmark}
+              showAllContent
+            />
+            <CommentList
+              feed={feed}
+              comments={comments}
+              onReply={handleCommentReply}
+              onLike={handleCommentActions.like}
+            />
           </ScrollView>
         </KeyboardLayout>
       </Layout>
@@ -162,3 +180,19 @@ export default function FeedDetailScreen({ navigation, route }: FeedDetailScreen
     </>
   );
 }
+
+export const SuspendedFeedDetailScreen = (props: FeedDetailScreenProps) => {
+  return (
+    <ErrorBoundary fallback={<></>}>
+      <Suspense
+        fallback={
+          <Layout showHeader disableBottomSafeArea onBack={() => props.navigation.goBack()}>
+            <FeedSkeleton />
+          </Layout>
+        }
+      >
+        <FeedDetailScreen {...props} />
+      </Suspense>
+    </ErrorBoundary>
+  );
+};
